tests/test_processing.py
import unittest
from unittest.mock import patch, mock_open
from chatbot.processing import ChatbotProcessor

class TestChatbotProcessor(unittest.TestCase):
    def setUp(self):
        intents_json = '{"intents": [{"tag": "greeting", "patterns": ["Hi", "Hello"], "responses": ["Hello!", "Hi there!"]}]}'
        words_list = ['hi', 'hello']
        classes_list = ['greeting']

        with patch('builtins.open', mock_open(read_data=intents_json)) as mock_file, \
             patch('pickle.load') as mock_pickle_load, \
             patch('chatbot.processing.load_model') as mock_load_model:

            # Setup mock return values
            mock_pickle_load.side_effect = [words_list, classes_list]
            mock_load_model.return_value = None  # Assuming the model's methods will be mocked as needed

            # Initialize ChatbotProcessor with mock dependencies
            self.processor = ChatbotProcessor('mock_model.h5', 'mock_intents.json', 'mock_words.pkl', 'mock_classes.pkl')

    def test_clean_up_sentence(self):
        test_sentence = "Hello, how are you?"
        cleaned_sentence = self.processor.clean_up_sentence(test_sentence)
        self.assertIsInstance(cleaned_sentence, list)  # Ensure it returns a list
        # Further tests can check for the correctness of tokenization and lemmatization

    @patch('chatbot.processing.ChatbotProcessor.bow')
    @patch('chatbot.processing.ChatbotProcessor.predict_class')
    def test_chatbot_response(self, mock_predict_class, mock_bow):
        # Setup mock return values for predict_class and bow
        mock_predict_class.return_value = [{"intent": "greeting", "probability": "0.9"}]
        mock_bow.return_value = [0, 1, 0]  # Example bow representation

        # The actual intents JSON structure would affect this test
        response = self.processor.chatbot_response("Hello")
        self.assertIsInstance(response, str)  # Verify response is a string
        # Add assertions based on the expected response for "Hello"

if __name__ == '__main__':
    unittest.main()
